Optimize travis build time
We introduce a number of optimizations to keep our build time short
even over a longer period of time:

* the size of the sccache cache is limited to 300M
* we clean the ~/.cargo/registry folder on the 1st of every month
* we no longer cache ~/.rustup in travis (downloading the rust
installation is roughly as fast as downloading from the cache but we
don't have to upload it in the end)

I also extracted the before_script part of the .travis.yml file into
dedicated shell scripts inside a .travis folder. This should be more
descriptive in what all that stuff is actually doing and help
maintainability.