Generalize Alice state machine for Bob
By adding another type parameter, we can generalize Alice's
state-machine in a way that it can be used for Bob as well. The new
type parameter S describes a type that can go into a SecretHash. For
Alice, this is the Secret type itself, whereas for Bob it is just the
SecretHash. The state-machine only uses the functionality of
converting this type into a SecretHash and can thus depend on the
generic Into conversion.

Alice needs the actual Secret again in her actions, where we have to
specify all the concrete types anyway. Here we only implement the
actions for states that carry a Secret.