Store pending responses from bam_api
This introduces a `PendingResponses` component which is used by Bob's
SwapRequestHandler to store pending responses. The http_api will use
this component to send the user's response to the swap request which
will result in Bob's state machine advancing to the next state AND the
response will be sent to Alice.