Use Option for CallContract member 'data'
Currently we use bytes.default() to generate an empty value for bytes array
`data` in the `CallContract` structure.  This results in the value `0x` being
used (I do not know the exact reason).  Rust has options for such a case.

This needs testing, I have no idea what will now be expressed by the `None` in
place of the original `0x` (FTR I do not know where the `Ox` is either :)

Fixes: #999