e2e: Increase maximum allowable transaction fee
Currently we use a maximum allowable bitcoin transaction fee of 5000
satoshi.  This number is arbitrary and it turns out it is not big
enough.

Recently we fixed a bug in e2e test creation code that was silently
preventing some tests from running, this bug was hiding the max fee
no big enough issue.

Increase maximum allowable transaction fee from 5000 to 50000.