Rename BehaviourInEvent to ProtocolInEvent
There is no such thing as a BehaviourInEvent because all events for
the behaviour are triggered through function calls _on_ the behaviour.
In our case, the behaviour only creates events that the ProtocolHandler
is supposed to handle. Libp2p even referres to this type as
ProtocolHandler::InEvent inside the NetworkBehaviour::poll method definition.