Move frame.rs up a directory
In preparation for renaming json/ to frame/ move json/frame.rs up a directory.
This makes sense because frame.rs defines `Frame` which is a top level BAM
construct.  With this in place and the directory rename we will be able to write
`bam::frame::Request` instead of `bam::json::Request`.