Remove integration tests
Currently we use integration tests because we hit the filesystem during
the tests AND we touch the config 'production' config file.  If we do
not touch the config file i.e., only test the config file writing on a
temporary file then we can safely move these integration tests to unit
tests.  The new tests will still be 'unit' tests because the unit under
testing will be the `read` and `write` methods of `File`.

In preparation for adding test functionality via unit tests remove the
integration tests, do this as a separate patch because we are going to
move to using the `directories` library before re-adding the tests.