Wait 1 second before ledger actions in tests
This hopefully eradicates bugs where one party performs a ledger
action before the other one has even had a chance to create a query
for it on btsieve (only applies to tests).

Look to increase the sleep time if this bug still happens.

Remove this once btsieve is able to look at blocks prior to the
creation of a query.