Use custom types in Queryable/Insertable structs
Currently we use `String` types for the fields (currently just swap_id)
in the structs that are written to and read from the database.  This
means we are not taking advantage of the Rust compile and the type
system.  We can do better than this.

Define and use a custom `SwapId` type for the swap_id struct fields
written to and from the Sqlite database.  This requires implementing the
`ToSql` and `FromSql` traits (from diesel).