Write strings to database for enums
Currently we write integers to the database to represent enums.  This
makes reading the sqlite file difficult but more importantly integers
make the database brittle because reordering the enum variants in code
will change the meaning of the data stored in the database.  We can use
strings to write to and from the database for enums fixing both these
issues.