Remove global variable
The code around the global block variable (and receipt) is wrong because
of how `include_str` works by embedding the file contents into the
binary.  This has the effect that the code just parses the string to
JSON only once and has no effect on the number of file reads (which is
already just a single read).

Remove the global variable and just use normal functions.  Currently the
functions are only called in a single place but when used by each of the
tests these functions will make the code obscure.  Add a note for the
next dev to implement an macro to solve this issue before using these
functions.  (This ticket is now out of sprint scope.)