Move role specific State out of spawner
Currently we spawn alice and bob for accept and decline actions and then
ignore decline within spawn.  This requires access to the `State` object
which is specific to alice/bob so spawn must know the role.  We do not
want spawn to know the role.

With this done we can move role specific State out of spawn().