Introduce high-level Accept and Decline messages for rfc003
Up until now, we had a high-level Request message that is split
into headers and body for the actual serialization. We didn't have
the same thing for Accept and Decline.

Most importantly, the id is missing from both of request bodies
because of the way our transport layer is currently implemented:
we have an open substream and hence, the accept and decline message
on the wire is scoped to only one request.

Out of this context, it is impossible to identify, where the
data in the accept or decline message belongs to. We do need this
data for storing the messages permanently on disk.

This patch therefore introduces high-level messages for accept
and decline that also track the SwapId.