db: Merge `SaveMessage<M>` and `Save` into `Save<T>`
Now that we have all the database saving implemented it is clear that the `Save`
trait (used for saving what was swap metadata) is a a specific case of the
generic `SaveMessage<M>` if 'message' was the more general 'anything'.

Merge the two traits together and remove reference to 'messages'.