Refactor deserialization and serialization of ledger and asset values
We make heavy use of serde's "from" and "into" features to explicitly
model the format we expect on the HTTP API and the format we send out.

This allows us to shift most of the error handling to serde, which
in turn gives us better error messages. They are not perfect but
about as good as it gets if we stick to deserializing JSON with serde.

We could potentially have better error messages by:

- either taking apart the JSON manually but that would be a huge
coding effort and cumbersome to maintain because we could not make
use of many features of serde.
- simplifying the JSON model to reduce the complexity in the
deserialization process.