Add data directory configuration option
Currently we have `--seed-file` for the seed file and a config option
for the database.  Both of these files default to the systems data
directory with `comit/` appended.  It makes more sense to configure
these two items in the same fashion.

Add configuration option

	[Data]
	dir = path/to/data/dir

Use this path with `cnd.sqlite` and `seed.pem` appended for the database
and seed file respectively.

Remove `--seed-file` command line option and the database configuration
option.