Encourage only Alice to derive the secret
Currently Bob can derive the secret from the seed that cnd starts up
with, this is incorrect.  This patch is an attempt to encourage only
Alice to derive the secret.  I say 'encourage' because via the facade
Bob can still derive the secret but with this applied we limit the
functionality to only deriving identities when doing the redeem action.