http_api: Use Facade as arg instead of trait bounds
Currently we pass the dependencies throughout the http_api module using
trait bounds.  This is super verbose.  Also since we now this is only
ever going to be an actual instance of the `Facade` and we are not
trying to limit the callers usage of the facade it adds no real benefit
using trait bounds.