Clean up Network code
Currently `main.rs` contains a bunch of networking code.  The root cause
of much of this is due to libp2p types.  If we wrap the libp2p swarm in
a struct then we can do a lot of cleaning up.

With this applied there is no more libp2p code in `main.rs` :)