Make Quickcheck code independent of the prod SECP context
Currently, we use a lazy-static SECP context in our library
to have access to this context at any time. This kind of global
access is a code-smell and should be avoided.

Unfortunately, our current bitcoin actions still depend on this
context being globally accessible.
With the transition to miniscript, this will hopefully go away and
we can then just pass the SECP context down to the HTTP controllers
through the facade and can ditch the lazy_static context (and the
dependency with it).

Removing the link between the current lazy_static context and the
quickcheck code is one step into this direction.