Fix action adding logic
There is a bug in the action adding logic, actions do net get added ever
if `on_fail == OnFail::OnAction`, this is incorrect.  We should only not
add the actions if the swap is in an error state AND `on_fail ==
OnFail::OnAction`.

Make the error checking code above uniform also.