Don't (ab)use the CreateInvoiceResponse to pass around data
Instead of using CreateInvoiceResponse everywhere, we now only
expose the data we actually need: invoice id and payment request.

This is enough for Alice to make a payment and for Bob to assert
that it was paid.

This patch also moves the assertion out of the Lightning wallet
class and into the Actor class where all the other assertions
reside.