Upgrade libp2p to version 0.16
Unfortunately, I had to trash all the tests in libp2p-comit because
I just couldn't get them to work. The e2e tests work though and we
are going to trash this communication protocol soon anyway. Hence,
in the interest of not wasting more time, I deleted them.

On the upside, libp2p 0.16 offers a lot of very nice things:

1. The TSubstream type variable is gone which reduces the overall
verbosity of the code.
2. The update to futures 0.3 and async/await is obviously nice.
3. We no longer need to print the errors of a ProtocolHandler
ourselves, this is now done in the Swarm (on level TRACE).