Inline refund/redeem logic
Currently we define and use helper functions for redeem/refund actions
for Ether and Erc20.  This is not strictly correct since it relies on
the fact that the log messages are the same.  This will change with the
protocol split Han/Herc20.  Lets fix this now to save time in the
future.