Remove the Into<LedgeKind> trait bound from Ledger
We are in the process of attempting to remove the `Ledger` trait bound
from various structs.  As a step towards that we can remove the
`Into<LedgerKind>` trait bound from within the `Ledger` trait.

Doing so requires manually implementing conversion from an
`rfc003::Request` to an `OutboundRequest` using concrete types.

Fixes: #2065