Remove Clone supertrait from Asset
We could have achieved this by cluttering the whole codebase with
`where: A: Clone` calls but that would kind of defeat the point.

Instead, we look at _why_ we needed to clone the asset in the first
place. Turns out, most of the usages can be replaced by references
up to a stage where the `HtlcParams` get passed into functions
with concrete types.

As soon as the compiler has type information available, we can
simply call `clone` if necessary without having to add further
trait bounds.