Add `identity` module
Our domain is cross chain.  We should have domain specific terms for
things in various chains.  We currently have the `asset` module and the
`ledger` module that do just this.

Add and `identity` module that defines an identity/address on a chain.
With this applied we can write

```
	let address = identity::Ethereum::new(...)
```
Instead of
```
	let address = ethereum::Address::new(...)
```
which is actually just a re-export of:
```
	let address = web3::types::Address::new(...)
```

Similarly we have `identity::Bitcoin` for the Bitcoin addresses.