Remove the `Asset` trait
Removing the `Asset` trait forces us to put `Ord` bounds in several
places because we have to compare the assets (expected vs actually
found in the HTLC).

It turns out that this is only required in very selective places
which is why I opted to remove it.

If we find this too verbose, we can add a trait again at some point.