Remove the Identity associated type from Ledger
In an effort to remove the `Ledger` trait first remove the associated
type `Identity`.  While doing this use the new `identity` modules.

Replace useages of `crate::ethereum::Address` with
`identity::Ethereum`.  We have to be a bit more careful with `PublicKey`
because we have a wrapper by the same name.