Add step to ensure network/chain id specified in config matches with connected node
Add a step during initialisation to test whether the chain/network id's of the
Ethereum/Bitcoin nodes that cnd has connected to matches the chain/network id's
specified in the config. Cnd will abort if the config and the actual chain/network
id does not match. Cnd will log a warning if it cannot make a request to the nodes.

Resolves: #2088