Don't clone blockchain connectors
Initially, we thought it was a good idea to make the traits
BlockByHash and LatestBlock take &mut self. It turns out that we
don't need this functionality for anything other than the tests.

Removing the mutable reference to self allows us to stop cloning
the connector and simply pass references into all the functions.