Simplify the walk_back_until algorithm
This patch improves the walk_back_until algorithm by removing the
duplication at the beginning the function. With some simple
re-orderings and re-assignments, we can remove a call to `clone`
on the block and remove the duplication in the `stop_condition`
check which is now named `should_stop_here` to be more descriptive.