Make config validation function generic and condense error handling
The Ethereum and Bitcoin node config validation functions were
replaced with a generic one. A trait was added to provide a
cleaner API to the caller and reduce the amount of imports
required to call the the validation function in main. Error
handling was condensed and cnd will crash if a unspecified error
occurs when trying to contact the blockchain node to validate
the config.