Add skeleton API function for controller/swarm interface
In order to be able to utilise the new communication protocol we need an
interface between the controller and the swarm.

Add an API function that allows the controller (acting as Alice) to call
the swarm (libp2p layer) and trigger execution of the new communication
protocols (by calling through to the underlying swarm) in order to
announce the created swap, send the known parameters, and receive the
remaining required swap parameters back from Bob.

Fixes: #2170