Wait for log files to exist before search for lines
In some cases, we ran into errors where the log reader tried to
read the file before it existed. Additionally, the design of
`LogReader` suggested that its instance can actually be reuse,
which is false. If the line we are looking for has been emitted
in the log file before we start looking for it, we never find it.

We circumvent both problems by removing the LogReader alltogether
and only exposing a single function.